requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
beautifulsoup4>=4.9.0
firebase-admin>=5.0.0
pymongo>=4.0.0
google-api-core>=2.0.0
python-binance>=1.0.16
pytz>=2022.1
matplotlib>=3.5.0
seaborn>=0.11.0
openai>=1.5.0
python-dotenv>=1.0.0 